Update searchResult_answer.php
Fixed code hanging off, but was unable to fix the links here (or in any other stream entries).

It throws unknown variable context/contentContainer. My best effort was adding '$contentContainer', before '/questionanswer/question/view/', which worked, but for some reason the resulting URL was domain.com/s/space-name/$contentContainer/questionanswer/question/view

So $contentContainer was getting doubled. The first time it was being rendered correctly and the second time it was rendered in plaintext. I'm not sure why it was rendered twice to begin with.